+++ CDJSVis/gui/infoDialogs.py
@@ -366,7 +366,7 @@ class DefectInfoWindow(QtGui.QDialog):
             radius = refState.specieCovalentRadius[refState.specie[index]] * self.filterList.displayOptions.atomScaleFactor
             
             # can do this because defect filter is always by itself
-            vacScaleSize = self.filterList.currentSettings[0].vacScaleSize
+            vacScaleSize = self.filterList.getCurrentFilterSettings()[0].vacScaleSize
             radius *= vacScaleSize * 2.0
             
             # highlighter
@@ -428,7 +428,7 @@ class DefectInfoWindow(QtGui.QDialog):
             radius = refState.specieCovalentRadius[refState.specie[vacIndex]] * self.filterList.displayOptions.atomScaleFactor
             
             # can do this because defect filter is always by itself
-            vacScaleSize = self.filterList.currentSettings[0].vacScaleSize
+            vacScaleSize = self.filterList.getCurrentFilterSettings()[0].vacScaleSize
             radius *= vacScaleSize * 2.0
             
             # highlight